Create models.py
The app has two models for now. A product which has title, description, slug, regular price, discount price, is_active, categories (which is a manytomanyfield), created_at, updated_at.

One Product can be a part of many Categories, and there could be subcategories of categories, therefore a self foreign key is introduced as a MPTT model for efficiency.